micro_service_with_gRPC/service_question_answer/requirements.txt
grpcio==1.70.0
grpcio-tools==1.70.0
protobuf==5.29.3
tensorflow==2.18.0
numpy==2.0.2
transformers==4.27.4